+++ app/src/main/java/com/jbrunoo/digitink/domain/repository/TicketRepository.kt
@@ -1,4 +1,4 @@
-package com.jbrunoo.digitink.domain
+package com.jbrunoo.digitink.domain.repository
 
 import com.jbrunoo.digitink.domain.model.Ticket
 import kotlinx.coroutines.flow.Flow
